ISPN-7900 Provide entrySet, values, keySet implementation for RemoteCache
* Refactor removable iterator to be more flexible
* Allows for remove to be more specific including value if needed
