+++ core/src/main/java/org/infinispan/stream/impl/local/KeyStreamSupplier.java
@@ -66,6 +66,6 @@ public class KeyStreamSupplier<K, V> implements AbstractLocalCacheStream.StreamS
 
    @Override
    public CloseableIterator<K> removableIterator(CloseableIterator<K> realIterator) {
-      return new RemovableCloseableIterator<>(realIterator, cache, Function.identity());
+      return new RemovableCloseableIterator<>(realIterator, cache::remove);
    }
 }
